requirements.txt
dnspython==2.7.0
prometheus_client==0.22.1
PyYAML==6.0.2